Finished refactoring sub components into seperate files.
* De-linted main `ReviewsAndRatings.jsx` file
